Add comment to cashOut saga
The button responsible for dispatching a cashOut action that triggers the cashOut saga is conditionally rendered based on the passing of all checks in the warningStack module, which includes a check on a valid account being connected. This error handler is here for the scenario where code isn't running as expected, which could be caused by any number of weird metamask shenanigans, browser differences, server-errors, DDOS attempts, etc. If any of these happen, we want to block the user from hitting our API to request a cashOut.